requirements.txt
# S1 minimal locked dependencies for local development and testing
colorama==0.4.6
iniconfig==2.3.0
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
Pillow>=10.0.0
PyYAML>=6.0
pytest==9.0.3
